+++ src/commands/hardis/work/save.ts
@@ -41,6 +41,7 @@ Example:
 
 \`\`\`yaml
 autoCleanTypes:
+  - checkPermissions
   - destructivechanges
   - datadotcom
   - minimizeProfiles