+++ CHANGELOG.md
@@ -7,7 +7,15 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 ## [4.12.3] 2023-11-18
 
 - Monitoring: Update default **package-skip-items.xml**
-- **hardis:diagnose:audittrail**: Add PerSetUnassign in not suspect monitored actions in Setup Audit Trail
+- **hardis:org:diagnose:audittrail** enhancements:
+  - Add PerSetUnassign in not suspect monitored actions in Setup Audit Trail
+  - Allow to append more allowed Setup Audit Trail sections & actions using `.sfdx-hardis.yml` property **monitoringAllowedSectionsActions**
+- **hardis:lint:access** enhancements:
+  - Output report file
+  - Send slack notification
+  - Add it by default in the monitoring commands
+- Doc
+  - Update contributing infos (use `sf plugins link`)
 
 ## [4.12.2] 2023-11-15